Updated Order with new print Order methods
Had to add a payment field to the Order class to be able to print the payment type. Not sure if this is going to effect the GUI or if changes will need to be made to it to update the field when placing an Order through the GUI. Let me know what you all think.